@@ -24,21 +24,19 @@
 #
 """Information Package manifests."""
 import os
-from pathlib import Path
+import pickle
+from pathlib import Path, PurePath
+from typing import Optional
 
 import lxml.etree as ET
-from eark_validator.mets import MetsFiles
-from eark_validator.model.metadata import FileEntry
 
-from eark_validator.utils import get_path
 from eark_validator.const import NO_PATH, NOT_DIR, NOT_FILE
-from eark_validator.model import (
-    Checksum,
-    ChecksumAlg,
-    Manifest,
-    ManifestEntry
-    )
+from eark_validator.mets import MetsFiles
+from eark_validator.model import Checksum, ChecksumAlg, Manifest, ManifestEntry
 from eark_validator.model.manifest import SourceType
+from eark_validator.model.metadata import FileEntry
+from eark_validator.utils import get_path
+
 
 class Checksummer:
     def __init__(self, algorithm: ChecksumAlg | str):
@@ -54,7 +52,7 @@ class Checksummer:
             raise FileNotFoundError(NO_PATH.format(path))
         if (not path.is_file()):
             raise ValueError(NOT_FILE.format(path))
-        implemenation = ChecksumAlg.get_implementation(self._algorithm)
+        implemenation: ChecksumAlg = ChecksumAlg.get_implementation(self._algorithm)
         with open(path, 'rb') as file:
             for chunk in iter(lambda: file.read(4096), b''):
                 implemenation.update(chunk)
@@ -68,17 +66,18 @@ class Checksummer:
 
 class ManifestEntries:
     @staticmethod
-    def from_file_path(path: Path, checksum_algorithm: ChecksumAlg | str=None) -> ManifestEntry:
+    def from_file_path(root: Path, entry_path: Path, checksum_algorithm: ChecksumAlg | str=None) -> ManifestEntry:
+        abs_path: Path = root.joinpath(entry_path).absolute()
         """Create a FileItem from a file path."""
-        if (not os.path.exists(path)):
-            raise FileNotFoundError(NO_PATH.format(path))
-        if (not os.path.isfile(path)):
-            raise ValueError('Path {} is not a file.'.format(path))
+        if (not os.path.exists(abs_path)):
+            raise FileNotFoundError(NO_PATH.format(abs_path))
+        if (not os.path.isfile(abs_path)):
+            raise ValueError('Path {} is not a file.'.format(abs_path))
         algorithm = checksum_algorithm if isinstance(checksum_algorithm, ChecksumAlg) else ChecksumAlg.from_string(checksum_algorithm)
-        checksums = [ Checksummer.from_file(path, algorithm) ] if checksum_algorithm else []
+        checksums = [ Checksummer.from_file(abs_path, algorithm) ] if checksum_algorithm else []
         return ManifestEntry.model_validate({
-            'path': path,
-            'size': os.path.getsize(path),
+            'path': entry_path,
+            'size': os.path.getsize(abs_path),
             'checksums': checksums
             })
 
@@ -93,28 +92,21 @@ class ManifestEntries:
 
 class Manifests:
     @classmethod
-    def validate_manifest(cls, path: Path, manifest: Manifest) -> tuple[bool, list[str]]:
+    def validate_manifest(cls, manifest: Manifest, alt_root: Optional[Path] = None) -> tuple[bool, list[str]]:
         """Check the integrity of the manifest."""
-        is_valid = True
-        issues = []
+        issues: list[str] = []
+        root = alt_root if alt_root else _resolve_manifest_root(manifest)
         for entry in manifest.entries:
-            abs_path = Path(os.path.join(path, entry.path))
+            abs_path = Path(os.path.join(root, entry.path))
             if not abs_path.is_file():
-                is_valid = False
-                issues.append('File {} is missing.'.format(entry.path))
+                issues.append('File {} is missing.'.format(abs_path))
                 continue
             if (entry.size != os.path.getsize(abs_path)):
                 issues.append('File {} manifest size {}, filesystem size {}.'.format(entry.path, entry.size, os.path.getsize(abs_path)))
-                is_valid = False
-            calced_checksum = Checksummer.from_file(abs_path, entry.checksum.algorithm)
-            if not entry.checksum == calced_checksum:
-                issues.append('File {} manifest checksum {}, calculated checksum {}.'.format(entry.path, entry.checksum, calced_checksum))
-                is_valid = False
-        return is_valid, issues
-
-    @staticmethod
-    def _relative_path(root_path: str, path: str) -> str:
-        return path if not os.path.isabs(path) else os.path.relpath(path, root_path)
+            check_issues: list[str] = _test_checksums(abs_path, entry.checksums)
+            if not bool(check_issues):
+                issues.extend(check_issues)
+        return (not bool(issues)), issues
 
     @staticmethod
     def from_source(source: Path | str, checksum_algorithm: ChecksumAlg=None) -> Manifest:
@@ -127,6 +119,18 @@ class Manifests:
             raise ValueError('Path {} is neither a file nor a directory.'.format(source))
 
     @staticmethod
+    def to_file(manifest: Manifest, path: Path | str) -> None:
+        path = get_path(path, False)
+        with open(path, 'wb') as file:
+            pickle.dump(manifest, file)
+
+    @staticmethod
+    def from_file(path: Path | str) -> Manifest:
+        path = get_path(path, False)
+        with open(path, 'rb') as file:
+            return pickle.load(file)
+
+    @staticmethod
     def from_directory(source: Path | str, checksum_algorithm: ChecksumAlg=None) -> Manifest:
         path = get_path(source, True)
         if (not path.is_dir()):
@@ -134,9 +138,11 @@ class Manifests:
         entries = []
         for subdir, dirs, files in os.walk(source):
             for file in files:
-                file_path = Path(os.path.join(subdir, file))
-                entries.append(ManifestEntries.from_file_path(file_path, checksum_algorithm=checksum_algorithm))
+                root = Path(os.path.join(subdir, file))
+                entry_path = root.relative_to(path)
+                entries.append(ManifestEntries.from_file_path(path, entry_path, checksum_algorithm=checksum_algorithm))
         return Manifest.model_validate({
+            'root': path,
             'source': SourceType.PACKAGE,
             'summary': None,
             'entries': entries
@@ -148,9 +154,26 @@ class Manifests:
         if (not path.is_file()):
             raise ValueError(NOT_FILE.format(source))
         mets_file = MetsFiles.from_file(path)
-        entries: list[ManifestEntry] = list(map(ManifestEntries.from_file_entry, mets_file.file_references))
+        entries: list[ManifestEntry] = list(map(ManifestEntries.from_file_entry, mets_file.file_entries))
         return Manifest.model_validate({
+            'root': path,
             'source': SourceType.METS,
             'summary': None,
             'entries': entries
             })
+
+def _test_checksums(path: Path, checksums: list[Checksum]) -> list[str]:
+    issues: list[str] = []
+    for checksum in checksums:
+        calced_checksum = Checksummer(checksum.algorithm).hash_file(path)
+        if not checksum == calced_checksum:
+            issues.append('File {} manifest checksum {}, calculated checksum {}.'.format(path, checksum.value, calced_checksum))
+    return issues
+
+def _resolve_manifest_root(manifest: Manifest) -> Path:
+    if manifest.source == SourceType.PACKAGE:
+        return manifest.root
+    elif manifest.source == SourceType.METS:
+        return manifest.root.parent
+    else:
+        raise ValueError('Unknown source type {}'.format(manifest.source))
