@@ -32,19 +32,30 @@ from lxml import etree
 from eark_validator.ipxml.schema import IP_SCHEMA
 from eark_validator.ipxml.namespaces import Namespaces
 from eark_validator.model.checksum import Checksum, ChecksumAlg
-from eark_validator.model.metadata import FileEntry, MetsFile
+from eark_validator.model.metadata import FileEntry, MetsFile, MetsRoot
+from eark_validator.model.validation_report import Location, Result
 from eark_validator.utils import get_path
 from eark_validator.const import NOT_FILE, NOT_VALID_FILE
 
 class MetsFiles():
     @staticmethod
-    def from_file(mets_file: Path | str):
+    def details_from_mets_root(namespaces: dict[str,str], root_element: etree.Element) -> MetsRoot:
+        return MetsRoot.model_validate({
+            'namespaces': namespaces,
+            'objid': root_element.get('OBJID', ''),
+            'label': root_element.get('LABEL', ''),
+            'type': root_element.get('TYPE', ''),
+            'profile': root_element.get('PROFILE', '')
+            })
+
+    @staticmethod
+    def from_file(mets_file: Path | str) -> MetsFile:
         path: Path = get_path(mets_file, True)
         if (not path.is_file()):
             raise ValueError(NOT_FILE.format(mets_file))
         ns: dict[str, str] = {}
         entries: list[FileEntry] = []
-        objid = label = pkg_type = othertype = contentinformationtype = profile = oaispackagetype = ''
+        othertype = contentinformationtype = oaispackagetype = ''
         try:
             parsed_mets = etree.iterparse(mets_file, events=['start', 'start-ns'])
             for event, element in parsed_mets:
@@ -54,34 +65,27 @@ class MetsFiles():
                     ns[prefix] = ns_uri
                 if event == 'start':
                     if element.tag == Namespaces.METS.qualify('mets'):
-                        objid = element.get('OBJID', '')
-                        label = element.get('LABEL', '')
-                        pkg_type = element.get('TYPE', '')
+                        mets_root: MetsRoot = MetsFiles.details_from_mets_root(ns, element)
                         othertype = element.get(Namespaces.CSIP.qualify('OTHERTYPE'), '')
                         contentinformationtype = element.get(Namespaces.CSIP.qualify('CONTENTINFORMATIONTYPE'), '')
-                        profile = element.get('PROFILE', '')
                     elif element.tag == Namespaces.METS.qualify('metsHdr'):
                         oaispackagetype = element.get(Namespaces.CSIP.qualify('OAISPACKAGETYPE'), '')
-                    elif element.tag == Namespaces.METS.qualify('file') | element.tag == Namespaces.METS.qualify('mdRef'):
+                    elif (element.tag == Namespaces.METS.qualify('file')) or (element.tag == Namespaces.METS.qualify('mdRef')):
                         entries.append(_parse_file_entry(element))
         except etree.XMLSyntaxError:
             raise ValueError(NOT_VALID_FILE.format(mets_file, 'XML'))
         return MetsFile.model_validate({
-            'default_ns': ns,
+            'root': mets_root,
             'oaispackagetype': oaispackagetype,
-            'objid': objid,
-            'label': label,
-            'type': pkg_type,
             'othertype': othertype,
             'contentinformationtype': contentinformationtype,
-            'profile': profile,
             'file_entries': entries
             })
 
 class MetsValidator():
     """Encapsulates METS schema validation."""
     def __init__(self, root: str):
-        self._validation_errors: List[str] = []
+        self._validation_errors: List[Result] = []
         self._package_root: str = root
         self._reps_mets: Dict[str , str] = {}
         self._file_refs: List[FileEntry] = []
@@ -91,7 +95,7 @@ class MetsValidator():
         return self._package_root
 
     @property
-    def validation_errors(self) -> List[str]:
+    def validation_errors(self) -> List[Result]:
         return self._validation_errors
 
     @property
@@ -105,6 +109,10 @@ class MetsValidator():
     @property
     def file_references(self) -> List[FileEntry]:
         return self._file_refs
+
+    @property
+    def is_valid(self) -> bool:
+        return len(self._validation_errors) == 0
 
     def get_mets_path(self, rep_name: str) -> str:
         return self._reps_mets[rep_name]
@@ -126,7 +134,14 @@ class MetsValidator():
             for event, element in parsed_mets:
                 self._process_element(element)
         except etree.XMLSyntaxError as synt_err:
-            self._validation_errors.append(synt_err)
+            self._validation_errors.append(
+                Result.model_validate({
+                    'rule_id': 'XML-1',
+                    'location': Location.model_validate({ 'context': synt_err.filename, 'test': str(synt_err.lineno), 'description': str(synt_err.offset) }),
+                    'message': 'File {} is not valid XML. {}'.format(mets, synt_err.msg),
+                    'severity': 'Error'
+                    })
+            )
         return len(self._validation_errors) == 0
 
     def _process_element(self, element: etree.Element) -> None:
